requirements.txt
asgiref==3.8.1
boto3==1.37.11
botocore==1.37.11
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
cssselect==1.2.0
cssutils==2.11.1
Django==5.1.7
django-cors-headers==4.7.0
django-extensions==3.2.3
django-storages==1.14.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
docutils==0.16
drf-yasg==1.21.10
et_xmlfile==2.0.0
idna==3.10
inflection==0.5.1
jmespath==1.0.1
lxml==5.3.1
MarkupSafe==3.0.2
more-itertools==10.6.0
openpyxl==3.1.5
packaging==24.2
premailer==3.10.0
pyasn1==0.6.1
pydotplus==2.0.2
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.7.2
s3transfer==0.11.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uuid==1.30
Werkzeug==3.1.3
yagmail==0.15.293
